OpenStack: Stop using old nova authentication and switch to keystoneauth1
OpenStack has released a keystoen auth library (keystoneauth1) that we
should leverage for getting access. This change switches the code from
the old way of authenticating via options passed directly to novaclient
to using sessions managed by keystoneauth.

We populate the argparser with the required options for authenticating,
except for the "os-project-name" and "os-project-id", as we will rely on
the configured projects and VO mappings (so the configured user should
be member of that project, or must have the correct rights).

As a side effect, we stop using nova's proxy to glance, and start using
the glance client directly.

(cherry picked from commit c1df0ad08b5aead06a037c536a59398215f0af64)